Refactor the NonParameterizedTest
